Configure separate web and node builds
