+++ package.json
@@ -2,7 +2,8 @@
   "name": "initai-js",
   "version": "0.0.7",
   "description": "A JavaScript library for building rich experiences on top of the Init.ai platform.",
-  "main": "dist/initai.js",
+  "main": "dist/node/initai.js",
+  "browser": "dist/web/initai.js",
   "author": "Init.ai Developers <shared@init.ai>",
   "license": "UNLICENSED",
   "repository": {
@@ -12,8 +13,10 @@
   "scripts": {
     "cleanDist": "rm -rf dist",
     "dev": "yarn cleanDist && node tasks/development",
-    "dev:linked": "nodemon --watch src --exec 'webpack --progress && cp dist/initai.js dist/initai.min.js'",
-    "build": "yarn cleanDist && webpack --progress",
+    "dev:linked": "nodemon --watch src --exec 'webpack --config=config/webpack.web.js && webpack --config=config/webpack.node.js'",
+    "build:web": "webpack --config=config/webpack.web.js",
+    "build:node": "webpack --config=config/webpack.node.js",
+    "build": "yarn cleanDist && yarn build:web && yarn build:node",
     "build:production": "./tasks/build.sh",
     "test": "./tasks/test.sh",
     "test:watch": "yarn test -- --verbose --watch",
@@ -65,6 +68,7 @@
     "moment": "^2.18.1",
     "nodemon": "^1.11.0",
     "normalize.css": "^7.0.0",
+    "object-assign-deep": "^0.3.1",
     "prettier": "^1.5.2",
     "react-dev-utils": "^3.0.2",
     "source-map-loader": "^0.2.1",
